// src/apis/types.service.ts
export interface ICityLatLong {
  lat: string;
  lon: string;
  display_name: string;
  class?: string;
  type: string;
}

export interface IImgurResult {
  data: IImgurData;
  success: boolean;
  status: number;
}

export interface IImgModel {
  image : string;
}

export interface IImgurData {
  link: string;
  id?: string;
  type?: string;
  width?: number;
  height?: number;
  size?: number;
  name?: string;
}

export interface IGoogleImgResult {
  knowledge_graph?: IImageInformation[];
  visual_matches?: IVisualMatches[];
}

export interface IImageInformation {
  title?: string;
  subtitle?: string;
  thumbnail?: string;
  link?: string;
  images?: IImageInfoList[];
}

export interface IImageInfoList {
  title?: string;
  source?: string;
  link?: string;
}

export interface IVisualMatches {
  position: string;
  title: string;
  link: string;
  thumbnail: string;
  source: string;
  source_icon: string;
}

export interface IGoogleResult {
  search_metadata?: ISeachMetaData;
  search_parameters?: ISearchParameter;
  knowledge_graph?: ISearchInformation;
  top_sights?: ITopSights;
  inline_images?: ILocationImage[];
  organic_results?: IResultInfo[];
  error?: string;
  lat?: number;
  long?: number;
}

export interface ITopSights {
  sights?: ISights[];
  show_more_link?: string;
}

export interface ISights {
  title: string;
  link: string;
  rating: number;
  reviews: number;
  thumbnail: string;
  description: string;
}

export interface ISeachMetaData {
  id: string;
  status: string;
  json_endpoint: string;
  created_at: string;
  processed_at: string;
  google_url: string;
  raw_html_file: string;
  total_time_taken: number;
}

export interface ISearchParameter {
  engine: string;
  q: string;
  gl: string;
  device: string;
}

export interface ISearchInformation {
  title?: string;
  type?: string;
  header_images?: IHeaderImageInfo[];
  description?: string;
  local_map?: ILocationMapInfo;
  weather?: string;
  population?: string;
  area_code?: string;
  climate?: string;
  postal_code?: string;
  region?: string;
  founded?: string;
  buttons?: ILocationButtons[];
}

export interface ILocationImage {
  link: string;
  source: string;
  thumbnail: string;
  original: string;
  title: string;
  source_name: string;
}

export interface IHeaderImageInfo {
  image: string;
  source: string;
}

export interface ILocationMapInfo {
  image: string;
  link: string;
}

export interface IResultInfo {
  title: string;
  link: string;
  thumbnail: string;
  snippet: string;
}

export interface ILocationButtons {
  text?: string;
  title?: string;
  link?: string;
  displayed_link?: string;
  snippet?: string;
  snippet_highlighted_words?: string[];
}
// ...
export interface IChatGPTCall {
  id: string;
  object: string;
  created: number;
  model: string;
  choices: IChatGPTResponse[];
}

export interface IChatGPTResponse {
  text: string;
  index: number;
  logprobs?:string;
  finish_reason: string;
}