// src/redux/master/master.types.ts
export interface IGetImgUrl {
  type: Types.MASTER_GET_IMG_URL;
  payload: IImgModel;
}

export interface IGetImgUrlSuccess {
  type: Types.MASTER_GET_IMG_URL_SUCCESS;
  payload: {ImgurResult: IImgurResult};
}

export interface IGetGoogleResult {
  type: Types.MASTER_GET_GOOGLE_RESULT;
  payload: {query: string, lang: string};
}

export interface IGetGoogleImgResult {
  type: Types.MASTER_GET_GOOGLE_IMG_RESULT;
  payload: {imglink: string, lang: string};
}
// ...
export interface IGetCityLatLong {
  type: Types.MASTER_GET_CITY_LATLONG;
  payload: {cityname: string};
}
// ...
export interface ISendMessagesToChatGPT {
  type: Types.MASTER_SEND_MESSAGES_TO_CHATGPT;
  payload: {prompt: string, max_tokens: number};
}